+++ packages/common/src/Decorators/inline-styles.js
@@ -1,6 +1,0 @@
-export const HEADING = {
-  NORMAL: 'inline-normal-text',
-  ONE: 'inline-header-one',
-  TWO: 'inline-header-two',
-  THREE: 'inline-header-three',
-};