+++ packages/editor/src/RichContentEditor/getStrategyByStyle.js
@@ -1,4 +1,0 @@
-export const getStrategyByStyle = style => (contentBlock, callback) => {
-  contentBlock.findStyleRanges(character => character.hasStyle(style),
-    (start, end) => callback(start, end));
-};